Moved decrease and increase to its own folder called scoring
