+++ assets/js/utils/scoring/decrease.js
